Coverity instructions for Travis CI
When pushing to this branch we would also execute coverity scans. Typically we would not want to do this very often. As of 11.09.2015 Coverity dictates :

https://scan.coverity.com/faq#frequency
Up to 12 builds per week, with a maximum of 3 builds per day, for projects with fewer than 100K lines of code
Up to   8 builds per week, with a maximum of 2 builds per day, for projects with 100K to 500K lines of code
Up to   4 builds per week, with a maximum of 1 build per day, for projects with 500K to 1 million lines of code
Up to   2 builds per week, with a maximum of 1 build per day, for projects with more than 1 million lines of code